ran_copilot_agentcore/src/requirements.txt
boto3
python-dotenv
strands-agents
pandas
numpy
